functions/requirements.txt
firebase_functions~=0.1.0
firebase_functions~=0.1.0
pandas~=2.1.2
google_cloud_pubsub~=2.18.4
aiohttp~=3.8.6
